[QLN-179] ref: move search results
